% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_buildTree}
\alias{rcpp_buildTree}
\title{Build a cached tree adjacency structure}
\usage{
rcpp_buildTree(edge, root)
}
\arguments{
\item{edge}{Integer matrix, the tree's edge matrix}

\item{root}{the node number of the root}
}
\value{
An external pointer to the cached structure
}
\description{
Build a cached tree adjacency structure
}
