% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_getMRCA}
\alias{rcpp_getMRCA}
\title{Fast MRCA using cached depth + parent lookup}
\usage{
rcpp_getMRCA(tree_ptr, a, b)
}
\arguments{
\item{tree_ptr}{External pointer from rcpp_buildTree()}

\item{a}{Integer node ID (e.g. tip index)}

\item{b}{Integer node ID (e.g. tip index)}
}
\value{
Integer node ID of the most recent common ancestor
}
\description{
Fast MRCA using cached depth + parent lookup
}
