% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_getNonDescendantsFast}
\alias{rcpp_getNonDescendantsFast}
\title{Get leaves that are NOT descendants using a cached tree structure}
\usage{
rcpp_getNonDescendantsFast(tree_ptr, node, nleaves)
}
\arguments{
\item{tree_ptr}{External pointer from rcpp_buildTree()}

\item{node}{Integer node ID}

\item{nleaves}{Integer number of leaves}
}
\value{
Integer vector of descendant node IDs
}
\description{
Get leaves that are NOT descendants using a cached tree structure
}
