% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_nodepathFast}
\alias{rcpp_nodepathFast}
\title{Get the path between two nodes in a tree}
\usage{
rcpp_nodepathFast(tree_ptr, from, to)
}
\arguments{
\item{tree_ptr}{External pointer from rcpp_buildTree()}

\item{from}{Node number at the start of the path}

\item{to}{Node number at the start of the path}
}
\description{
Get the path between two nodes in a tree
}
