% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateSubstitutions}
\alias{simulateSubstitutions}
\title{Simulate a substitution count down each branch of a tree. First, the branch rates are sampled
from a correlated, uncorrelated, or OU process that can be dependent on traits.
Then the number of substitutions along each branch is sampled from a Poisson distribution.}
\usage{
simulateSubstitutions(
  time.tree,
  nu = 0.5,
  sigma = 0.5,
  beta = 0,
  theta = 1,
  traits = NULL,
  number.of.subst = 1000,
  method = c("AC", "UCLN", "TD")
)
}
\arguments{
\item{time.tree}{a binary rooted tree (phylo object)}

\item{nu}{variance scale of autocorrelated clock or TD}

\item{sigma}{standard deviation of UCLN}

\item{beta}{effect size of traits on rates (if traits is not NULL)}

\item{theta}{theta term for OU process in TD}

\item{traits}{one trait per node; set to NULL if rates are conditionally independent of traits}

\item{number.of.subst}{expected number of substitutions per unit of time}

\item{method}{clock model may be uncorrelated lognormal (UCLN),
autocorrelated lognormal (AC), or trait dependent (TD)}
}
\value{
Two trees, with branch lengths set to either subst. rates or counts, and two vectors,
one of true branch rates and one of estimated branch rates
(i.e., count divided by time, which can evaluate to zero)
}
\description{
Simulate a substitution count down each branch of a tree. First, the branch rates are sampled
from a correlated, uncorrelated, or OU process that can be dependent on traits.
Then the number of substitutions along each branch is sampled from a Poisson distribution.
}
\examples{
# Sample a birth-death tree with 50 taxa
time.tree <- ape::rphylo(birth=10, death=5, n=50)

# Simulate traits down the tree under Brownian motion
traits <- simulateTrait(time.tree)

# Simulate substitutions that have a positive association with traits, 
sim.result <- simulateSubstitutions(time.tree=time.tree, 
																		beta=1, 
																		theta=10, 
																		sigma=0.5, 
																		traits=traits, 
																		method="TD", 
																		number.of.subst=10000)
subst.tree <- sim.result$subst.tree.est
node.rates <- sim.result$node.rates.est
plot(subst.tree)
axis(1)

# Simulate substitutions under a Brownian autocorrelated clock, 
# with no association with traits
sim.result.brownian <- simulateSubstitutions(time.tree=time.tree, 
																							sigma=0.5, 
																							method="AC", 
																							number.of.subst=10000)
}
