% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{Build a distance matrix from a tree (time tree or substitution tree)}
\usage{
getDistanceMatrix(tree)
}
\arguments{
\item{tree}{a tree (phylo object)}
}
\value{
A distance matrix, where dij is the mean distance from leaf i and j to their ancestor
}
\description{
Build a distance matrix from a tree (time tree or substitution tree)
}
