% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylowise-package.R
\docType{package}
\name{phylowise-package}
\alias{phylowise}
\alias{phylowise-package}
\title{phylowise: Phylogenetic Pairwise Contrasts}
\description{
A phylogenetic comparative method for finding associations between biological traits and molecular evolutionary rates. The method samples pairs from a phylogeny such that each pair has non-overlapping edge paths, and can therefore be treated as statistically independent observations. Linear regression is performed on the pair contrasts. This approach is similar to phylogenetically independent contrasts (PIC) but without reconstructing the traits at internal nodes, and is better suited for finding trait-rate associations than phylogenetic generalised least squares (PGLS). Refer to Douglas and Bromham (2026) \doi{10.64898/2026.08.13.744736} for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jordandouglas/phylowise}
}

}
\author{
\strong{Maintainer}: Jordan Douglas \email{jordan.douglas@auckland.ac.nz}

Authors:
\itemize{
  \item Jordan Douglas \email{jordan.douglas@auckland.ac.nz}
  \item Lindell Bromham
}

}
\keyword{internal}
