% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylowise-package.R
\docType{package}
\name{phylowise-package}
\alias{phylowise}
\alias{phylowise-package}
\title{phylowise: Phylogenetic Pairwise Contrasts for Identifying Drivers of Molecular Rate Variation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jordandouglas/phylowise}
}

}
\author{
\strong{Maintainer}: Jordan Douglas \email{jordan.douglas@auckland.ac.nz}

Authors:
\itemize{
  \item Jordan Douglas \email{jordan.douglas@auckland.ac.nz}
}

}
\keyword{internal}
