% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonSampling.R
\name{plotPairs}
\alias{plotPairs}
\title{Plot a set of sampled pairs onto a tree}
\usage{
plotPairs(
  tree,
  pairs.df,
  edge.col = "red",
  show.tip.label = F,
  edge.width = 1,
  edge.width.pairs = 3,
  label.cex = 1
)
}
\arguments{
\item{tree}{the binary rooted tree used to get samples}

\item{pairs.df}{a data frame of pairs, obtained using phylowise::sampleTaxonPairs}

\item{edge.col}{the pairs will be highlighted in this colour}

\item{show.tip.label}{display tip labels on the tree?}

\item{edge.width}{edge line width of all branches on the tree}

\item{edge.width.pairs}{edge line width of paired branches}

\item{label.cex}{tree tip label font size, if show.tip.label=TRUE
XXX TODO EXAMPLES XXX
XXX
XXX
XXX
XXX
XXX
XXX
XXX}
}
\description{
Plot a set of sampled pairs onto a tree
}
