% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{PPC.test}
\alias{PPC.test}
\title{Phylogenetic pairwise contrast test on a series of taxon pairs, using ordinary
least squares and Bayesian model averaging (BMA). This method does not do multivariate
regression, but it returns the data frame that can be used for it.}
\usage{
PPC.test(
  pairs.df,
  standardise = 2,
  logY = TRUE,
  extreme.value.threshold = 0,
  prior.weight = 0.5,
  OR = 1000,
  epsilon = 1e-06
)
}
\arguments{
\item{pairs.df}{data frame of taxon pairs}

\item{standardise}{should we standardise the evolutionary distance? 0 for no, 1 for standardise by sqrt(distance),
2 for standardise by sqrt(mrca age), 3 for standarise by mrca age. default and recommended setting: 2}

\item{logY}{should we take the logarithm of the evolutionary distance? default and recommended setting: true}

\item{extreme.value.threshold}{remove any observations more than this many standard deviations away from
the mean (distance); set to zero for no extreme value removal}

\item{prior.weight}{trait inclusion prior probbaility for Bayesian model averaging (see BMA::bic.glm)}

\item{OR}{Occam's window for Bayesian model averaging (see BMA::bic.glm)}

\item{epsilon}{precision of BMA probabilities for Bayes factor calculation: p\if{html}{\out{<epsilon and p>}}1-epsilon
will be respectively set to epsilon and 1-epsilon to avoid NaN calculations}
}
\value{
Vector of p-values and Pearson correlations (one element per trait), a data frame of
standardised and logged datapoints for doing regression, and posterior probabilities /
Bayes factors from BMA analyses
}
\description{
Phylogenetic pairwise contrast test on a series of taxon pairs, using ordinary
least squares and Bayesian model averaging (BMA). This method does not do multivariate
regression, but it returns the data frame that can be used for it.
}
\examples{
# Sample a birth-death tree with 100 taxa
time.tree <- ape::rphylo(birth=10, death=5, n=100)
ntips = length(time.tree$tip.label)

# Simulate traits down the tree under Brownian motion
traits1 <- simulateTrait(time.tree)
traits1.leaf <- traits1[1:ntips]

# Simulate substitutions that have a positive association with traits, 
sim.result <- simulateSubstitutions(time.tree=time.tree, 
									beta=2, 
									theta=10, 
									traits=traits1, 
									method="TD", 
									number.of.subst=10000)
subst.tree <- sim.result$subst.tree.est

# Sample taxon pairs within a window of (0.01, 0.2) time units
pairs.df <- sampleTaxonPairs(subst.tree=subst.tree, 
							covariate=traits1.leaf, 
							window.tree=time.tree, 
							dist.min=0.01, dist.max=0.2)

# Perform linear regression on the pairs
ppc <- PPC.test(pairs.df)
p.value <- ppc$p.var

# Plot the pairs. Depending on the random seed, we should usually
#  see a clear positive trend under these parameters
data.df <- ppc$data.df
plot(data.df$trait, data.df$distance.response, xlab="Trait contrast", ylab="Distance contrast")
}
