% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{readBeastTrees}
\alias{readBeastTrees}
\title{Read a single summary tree or a posterior distribution of annotated nexus trees generated by the
BEAST software suites, and apply burn-in.  Built on top of the phylotate::read_annotated function.}
\usage{
readBeastTrees(nexus.file, burnin = 0.1)
}
\arguments{
\item{nexus.file}{a nexus file}

\item{burnin}{proportion of burn-in to apply to the file (default: 10\%)}
}
\value{
A list of trees
}
\description{
Read a single summary tree or a posterior distribution of annotated nexus trees generated by the
BEAST software suites, and apply burn-in.  Built on top of the phylotate::read_annotated function.
}
\examples{
\dontrun{
trees = readBeastTrees("species.trees", burnin=0.2)
tree1 = trees[[1]]
plot(tree1)
print(tree1$node.comment)
}
}
