% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonSampling.R
\name{sampleTaxonPairs}
\alias{sampleTaxonPairs}
\title{Sample a set of taxon pairs from the tree. These taxa will have non-overlapping edges between their paths and can this be treated as statistcially independent.
The algorithm iteratively searches for two random taxa that satisfy sampling requirements until there are no more valid pairs. Each pair must descend from an MRCA
with dist.min <= tMRCA <=  dist.max. c++ is used to speed up the runtime of this code.}
\usage{
sampleTaxonPairs(
  subst.tree,
  covariate,
  dist.min,
  dist.max = Inf,
  response = NULL,
  nested = TRUE,
  maximise = TRUE,
  youngest = FALSE,
  window.tree = subst.tree,
  distance.matrix = NULL,
  verbose = FALSE
)
}
\arguments{
\item{subst.tree}{a binary rooted tree, with branch lengths in units of change (phylo object)}

\item{covariate}{a data frame of traits at the tips of the tree (rows are taxa, columns are traits); rows should be in the same order as tips in the tree}

\item{dist.min}{minimum distance that two tips must be apart from their ancestor (mean of both distances), on window.tree}

\item{dist.max}{maximum distance that two tips must be apart from their ancestor (mean of both distances), on window.tree}

\item{response}{provide a vector as a response trait instead of genetic distances in the tree (optional)}

\item{nested}{can taxon pairs be nested with each other?}

\item{maximise}{should we maximise the trait difference?}

\item{youngest}{take the youngest pair at each step (and therefore increase the number of pairs)?}

\item{window.tree}{tree that is the basis for building the distance matrix, if it is not provided; should have same taxa as 'tree'}

\item{distance.matrix}{provide an n x n distance matrix rather than recalculate from scratch}

\item{verbose}{print some statements along the way}
}
\value{
A data frame, where each row is a pair of taxa.
}
\description{
Sample a set of taxon pairs from the tree. These taxa will have non-overlapping edges between their paths and can this be treated as statistcially independent.
The algorithm iteratively searches for two random taxa that satisfy sampling requirements until there are no more valid pairs. Each pair must descend from an MRCA
with dist.min <= tMRCA <=  dist.max. c++ is used to speed up the runtime of this code.
}
