% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateSubstitutions}
\alias{simulateSubstitutions}
\title{Simulate a substitution count down each branch of a tree. First, the branch rates are sampled from a correlated, uncorrelated, or OU process that can be dependent on traits.
Then the number of substitutions along each branch is sampled from a Poisson distribution.}
\usage{
simulateSubstitutions(
  time.tree,
  nu = 0.5,
  sigma = 0.5,
  beta = 0.5,
  theta = 1,
  traits = NULL,
  number.of.subst = 1000,
  method = c("AC", "UCLN", "TD")
)
}
\arguments{
\item{time.tree}{a binary rooted tree (phylo object)}

\item{nu}{variance scale of autocorrelated clock or TD}

\item{sigma}{standard deviation of UCLN}

\item{beta}{effect size of traits on rates (if traits is not NULL)}

\item{theta}{theta term for OU process in TD}

\item{traits}{one trait per node; leave as NULL if rates are conditionally independent of traits}

\item{number.of.subst}{expected number of substitutions per unit of time}

\item{method}{clock model may be uncorrelated lognormal (UCLN), autocorrelated lognormal (AC), or trait dependent (TD)}
}
\value{
Two trees, with branch lengths set to either subst. rates or counts, and two vectors, one of true branch rates and one of estimated branch rates (i.e., count divided by time, which can evaluate to zero)
XXX TODO EXAMPLES XXX
XXX
XXX
XXX
XXX
XXX
XXX
XXX
}
\description{
Simulate a substitution count down each branch of a tree. First, the branch rates are sampled from a correlated, uncorrelated, or OU process that can be dependent on traits.
Then the number of substitutions along each branch is sampled from a Poisson distribution.
}
