% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateTrait}
\alias{simulateTrait}
\title{Simulate a trait down a tree using Brownian motion (BM)}
\usage{
simulateTrait(time.tree, sigma = 1, root.value = 0)
}
\arguments{
\item{time.tree}{a binary rooted tree (phylo object)}

\item{sigma}{standard deviation for BM}

\item{root.value}{trait value at the root of the tree}
}
\value{
A vector of traits, one element for each node in the tree, ordered by node number
XXX TODO EXAMPLES XXX
XXX
XXX
XXX
XXX
XXX
XXX
XXX
}
\description{
Simulate a trait down a tree using Brownian motion (BM)
}
